backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.9.1
httpx==0.27.2
python-multipart==0.0.12
pydantic==2.9.2
openai==1.51.0
python-jose[cryptography]==3.3.0
pandas==2.2.3
